+++ apps/betterangels-backend/shelters/admin.py
@@ -1,12 +1,11 @@
-from typing import Any, Optional, Type, TypeVar, cast
+from typing import Type, TypeVar, cast
 
 from common.admin import AttachmentInline
 from common.managers import AttachmentQuerySet
-from common.models import Attachment
 from django import forms
 from django.contrib import admin
 from django.db import models
-from django.forms import CheckboxSelectMultiple, ModelForm, SelectMultiple, TimeInput
+from django.forms import CheckboxSelectMultiple, SelectMultiple, TimeInput
 from django.http import HttpRequest
 
 from .enums import (
@@ -25,9 +24,9 @@ from .enums import (
     SleepingChoices,
     StorageChoices,
 )
+from .forms import BaseAttachmentForm, HeroForm
 from .models import (
     Accessibility,
-    Attachment,
     CareerService,
     City,
     EntryRequirement,
@@ -45,27 +44,7 @@ from .models import (
 )
 
 
-class BaseAttachmentForm(ModelForm):
-    def __init__(self, *args: Any, **kwargs: Any):
-        super().__init__(*args, **kwargs)
-        self.initial_file = self.instance.file if self.instance.pk else None
-
-    def save(self, commit: bool = True) -> Attachment:
-        new_attachment: Attachment = super().save(commit=commit)
-
-        # If there was an initial file and a new file has been uploaded
-        if self.initial_file and self.initial_file != new_attachment.file:
-            self.initial_file.delete(save=False)
-
-        return new_attachment
-
-
 class HeroInine(AttachmentInline):
-    class HeroForm(BaseAttachmentForm):
-        def save(self, commit: bool = True) -> Attachment:
-            self.instance.namespace = "hero_image"
-            return super().save(commit=commit)
-
     form = HeroForm
     max_num = 1
     verbose_name = "Hero Image"
